refactor: collapse AuthorityBaseCqlRepository into AuthorityCqlRepository
The Base layer existed to share CQL query logic between AuthorityRepository
and the now-deleted AuthorityArchiveRepository. With only one consumer left,
the four-class hierarchy is unnecessary.

- Delete AuthorityBaseCqlRepository interface
- Delete AuthorityBaseCqlRepositoryImpl class
- AuthorityCqlRepository now declares all four CQL methods directly
- AuthorityCqlRepositoryImpl now contains the full implementation
