+++ src/main/java/org/folio/entlinks/domain/repository/AuthorityCqlRepository.java
@@ -1,5 +1,17 @@
 package org.folio.entlinks.domain.repository;
 
-public interface AuthorityCqlRepository extends AuthorityBaseCqlRepository {
+import java.util.UUID;
+import org.folio.entlinks.domain.entity.Authority;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 
+public interface AuthorityCqlRepository {
+
+  Page<Authority> findByCql(String cql, Pageable pageable);
+
+  Page<UUID> findIdsByCql(String cqlQuery, Pageable pageable);
+
+  Page<Authority> findDeletedByCql(String cql, Pageable pageable);
+
+  Page<UUID> findDeletedIdsByCql(String cqlQuery, Pageable pageable);
 }
